The scrolling seems to be working.  There are several questions though.
If I remove the 1px border, for some reason the image get's moved and
starts at the top of the space-below.
If I change the left and right margin from 10% to 20% on .top-bar and
.main-picture, then the bottom bar doesn't start at the correct place
anymore.

I wanted for this to work no matter how you scale the window, but I'm
starting to think I might want to keep the image a certain size and figure
out the scrolling around that.