Working on a prototype for a word game.
Right now I'm trying to get the board built with flex box so I can make
sure that it can be resized.  I have the board being generated by
javascript since I don't konw how big it will eventually be.
I'm trying to figure out how to make the tiles square and span so that
they fill the div that contains them.

My laptop is about to die, so I need to check this in and probably go to
sleep.