Decided to make another prototype of a game that allows you to expand the board as you create your place your letters.
I also did a major refactor putting all the functions and variables
dealing with the board into an object.

There are still several refactors I need to do like move more
functionality into the move cursor functions so that the key handlers are
more simple and all the logic is in the board object.  It felt like a good
delineation between the board object and the key handlers.