Updated expandingLetterBoard to use css variables so that I'm not querying every tile and setting properties on each.  Now I'll just update the css and all the tiles will be updated.
I'm still having trouble when I add more y tiles than x tiles.  It sets
the height correctly, but doesn't shrink the width.  It seems like the
tiles are set somehow to fill the row instead of shrinking and centering
in the row.  Trying to figure out what to do about that.