Started working on the letter bank.  A player will only be allowed to type in letters that are in the letter bank.  Each letter you type, will remove it from the letter from the letter bank.  I still need to move the letter back into the letter bank when using the backspace button.  I'm thinking of implementing the delete button to remove letters in the opposite direction from the backspace.
I also grabbed a word.txt file that is a random dictionary that I found on
github.
I was also able to pull the 1,000 most popular words from the xkcd.com A
Thing Explainer word checker.  He had them as a string that was stored in
the window object.
I'm going to use those files to figure out the distribution of letters to
figure out how many of each to include in the total list of possible
letters.