I added an object that can simulate what the server does.  Right now it just shuffles letters and distributes them according to how many players there are.  It's defaulted to 4 players.  Right now I'm sticking to 100 letters since I'm imagining at more 5 players at a time right now.  I'm thinking that if there are more players, then I would go to 200.  I want to keep the idea of distributing a pile of letters evenly among the players.
The next step is doing the delete keyhandler and the enter key handler.  I
should add a button that does the same thing.  I have a feeling that using
the enter button to check the board would have people spamming the server.
I guess the better thing to do would be to have enter check the board
locally and then if it is completely filled out, then signal the server.
The server can then signal all the players to get two more letters.