Added the delete button for each new element added to the profile.  Right now I'm using a class for the input element and the button element to know what elements to delete.
I'm starting to think I should create a div and put the input and delete
button into it and then just delete the div.  That way I only have to
delete one element and I can use an id since there's only one thing to
delete.  It also might be easier to style it since they are already
together.