This was suprisingly hard.  I finally got the styling of the phone input correct.  I wanted the phone number and the phone number type to be side by side as if it's one element.  and the delete button connected to it.
I originally thought that I could add a new row and use the col width to
split it to where I wanted, but that proved to not work.  I could put the
select inside a div and get them to be side by side, but then the drop
down was unstyled and looked bad.  Then if I tried to set the select with
the class form-control, I would lose all of the options and it would only
show the down arrow.

I finally got it working by adding a 0 width span between the inputs and
then making the left border of the select input 0px.

Now I need to figure out how to make the + buttons look good since they
span the entire width of the form now.  I also need to figure out a way to
inform the user which element the + is for.

logTime = 3:00