Just finished working through the code review from Craig Jennings. Used the class keyword instead of creating an "object" which worked, but was very unconventional.
I was also trying to create member variables, but didn't see a great way
to do that.  I ended up calling this.<variable> in the constructor and it
seemed to create a member variable with the class without having to do
something like:
class MyClass {}
MyClass.foo = 'bar';

There was some other small bugs that I cleared up.