Updated the mongoDB model to use Date instead of String for the birthday.
After going through more of the LearnNode class, I think I'm going to
calculate age when presenting data on the client side.  I could hava a
script that runs on the server side periodically, but then I could be
updating a lot of profiles that don't need ot be.  Also, if I do this,
then I can remove the field from the model and offload the processing to the
client's computer